feat: web3 provider base class
