+++ src/ape_geth/providers.py
@@ -1,6 +1,6 @@
 import shutil
 from pathlib import Path
-from typing import Dict, Iterator, Mapping, Optional, Union
+from typing import Dict, Mapping, Optional, Union
 from urllib.parse import urlparse
 
 from eth_utils import to_wei
@@ -18,7 +18,7 @@ from web3.types import NodeInfo
 
 from ape.api import ReceiptAPI, TransactionAPI
 from ape.api.config import ConfigItem
-from ape.api.providers import ProviderAPI
+from ape.api.providers import Web3Provider
 from ape.exceptions import (
     ContractLogicError,
     OutOfGasError,
@@ -140,8 +140,7 @@ class GethNotInstalledError(ConnectionError):
         )
 
 
-class GethProvider(ProviderAPI):
-    _web3: Web3 = None  # type: ignore
+class GethProvider(Web3Provider):
     _geth: Optional[EphemeralGeth] = None
 
     @property
@@ -150,6 +149,7 @@ class GethProvider(ProviderAPI):
 
     def connect(self):
         self._web3 = Web3(HTTPProvider(self.uri))
+        self._web3.eth.set_gas_price_strategy(rpc_gas_price_strategy)
 
         # Try to start an ephemeral geth process if no provider is running.
         if not self._web3.isConnected():
@@ -178,8 +178,6 @@ class GethProvider(ProviderAPI):
                 self._geth.disconnect()
                 raise ConnectionError("Unable to connect to locally running geth.")
 
-        self._web3.eth.set_gas_price_strategy(rpc_gas_price_strategy)
-
         def is_poa() -> bool:
             node_info: Mapping = self._node_info or {}
             chain_config = extract_nested_value(node_info, "protocols", "eth", "config")
@@ -200,13 +198,8 @@ class GethProvider(ProviderAPI):
             self._geth.disconnect()
             self._geth = None
 
-        # Must happen after geth.disconnect()
+        # Note: this must happen after geth.disconnect()
         self._web3 = None  # type: ignore
-
-    def update_settings(self, new_settings: dict):
-        self.disconnect()
-        self.provider_settings.update(new_settings)
-        self.connect()
 
     def estimate_gas_cost(self, txn: TransactionAPI) -> int:
         """
@@ -215,8 +208,7 @@ class GethProvider(ProviderAPI):
         The transaction will not be added to the blockchain.
         """
         try:
-            txn_dict = txn.as_dict()
-            return self._web3.eth.estimate_gas(txn_dict)  # type: ignore
+            return super().estimate_gas_cost(txn)
         except ValueError as err:
             tx_error = _get_vm_error(err)
 
@@ -229,21 +221,6 @@ class GethProvider(ProviderAPI):
             raise TransactionError(base_err=tx_error, message=message) from err
 
     @property
-    def chain_id(self) -> int:
-        """
-        Returns the currently configured chain ID,
-        a value used in replay-protected transaction signing as introduced by EIP-155.
-        """
-        return self._web3.eth.chain_id
-
-    @property
-    def gas_price(self):
-        """
-        Returns the current price per gas in wei.
-        """
-        return self._web3.eth.generate_gas_price()
-
-    @property
     def _node_info(self) -> Optional[NodeInfo]:
         try:
             return self._web3.geth.admin.node_info()
@@ -251,62 +228,20 @@ class GethProvider(ProviderAPI):
             # Unsupported API in user's geth.
             return None
 
-    def get_nonce(self, address: str) -> int:
-        """q
-        Returns the number of transactions sent from an address.
-        """
-        return self._web3.eth.get_transaction_count(address)  # type: ignore
-
-    def get_balance(self, address: str) -> int:
-        """
-        Returns the balance of the account of a given address.
-        """
-        return self._web3.eth.get_balance(address)  # type: ignore
-
-    def get_code(self, address: str) -> bytes:
-        """
-        Returns code at a given address.
-        """
-        return self._web3.eth.get_code(address)  # type: ignore
-
-    def send_call(self, txn: TransactionAPI) -> bytes:
-        """
-        Executes a new message call immediately without creating a
-        transaction on the block chain.
-        """
-        return self._web3.eth.call(txn.as_dict())
-
-    def get_transaction(self, txn_hash: str) -> ReceiptAPI:
-        """
-        Returns the information about a transaction requested by transaction hash.
-        """
-        # TODO: Work on API that let's you work with ReceiptAPI and re-send transactions
-        receipt = self._web3.eth.wait_for_transaction_receipt(txn_hash)  # type: ignore
-        txn = self._web3.eth.get_transaction(txn_hash)  # type: ignore
-        return self.network.ecosystem.receipt_class.decode({**txn, **receipt})
-
     def send_transaction(self, txn: TransactionAPI) -> ReceiptAPI:
         """
         Creates a new message call transaction or a contract creation
         for signed transactions.
         """
         try:
-            txn_hash = self._web3.eth.send_raw_transaction(txn.encode())
+            receipt = super().send_transaction(txn)
         except ValueError as err:
             raise _get_vm_error(err) from err
-
-        receipt = self.get_transaction(txn_hash.hex())
 
         if txn.gas_limit is not None and receipt.ran_out_of_gas(txn.gas_limit):
             raise OutOfGasError()
 
         return receipt
-
-    def get_events(self, **filter_params) -> Iterator[dict]:
-        """
-        Returns an array of all logs matching a given set of filter parameters.
-        """
-        return iter(self._web3.eth.get_logs(filter_params))  # type: ignore
 
 
 def _get_vm_error(web3_value_error: ValueError) -> TransactionError:
